+++ Changelog.md
@@ -5,6 +5,7 @@
 - Removed duplicate Proclamation Day for Latvia, by @Daglina (#686).
 - New calendar: Added Kazakhstan calendar by @atj01 (#688).
 - New calendar: Added Georgia (country) calendar by @atj01 (#687).
+- Added conditional holidays on 26th December and 2nd January in Neuchatel (Switzerland) (#697).
 
 ## v16.2.0 (2021-12-10)